refactor: rename component.
