+++ packages/map-template/src/components/MapWrapper/GoogleMapsMap/GoogleMapsMap.scss
