+++ packages/map-template/src/components/MapWrapper/GoogleMapsMap/GooglePlacesHandler.js
