+++ packages/map-template/src/components/MapWrapper/MapboxMap/MapboxMap.jsx
