+++ packages/map-template/src/components/MapWrapper/MapboxMap/MapboxMap.scss
