+++ packages/map-template/src/components/MapWrapper/MapboxMap/MapboxPlacesHandler.js
