+++ packages/map-template/src/components/MapWrapper/MapboxMap/ViewmodeSwitch/ViewModeSwitch.jsx
